darlingCms_0.1_dev|core|classes: Refactored the MySqlPermissionCrud class: - The MySqlPermission class now implements the AMySqlPermissionCrud abstract class. - Removed methods now defined in parent class. - The MySqlPermissionCrud class now notifies it's observers whenever a permission is updated or deleted.

+++ core/classes/crud/MySqlPermissionCrud.php
@@ -1,18 +1,14 @@
 <?php
 /**
- * Created by PhpStorm.
- * User: sevidmusic
+ * Created by Sevi Donnelly Foreman.
  * Date: 2018-12-21
  * Time: 23:23
  */
 
 namespace DarlingCms\classes\crud;
 
-
-use DarlingCms\abstractions\crud\AMySqlQueryCrud;
-use DarlingCms\classes\database\SQL\MySqlQuery;
+use DarlingCms\abstractions\crud\AMySqlPermissionCrud;
 use DarlingCms\classes\privilege\Permission;
-use DarlingCms\interfaces\crud\IActionCrud;
 use DarlingCms\interfaces\crud\IPermissionCrud;
 use DarlingCms\interfaces\privilege\IPermission;
 
@@ -22,47 +18,8 @@ use DarlingCms\interfaces\privilege\IPermission;
  * to perform CRUD operations on Permission data in a MySql database.
  * @package DarlingCms\classes\crud
  */
-class MySqlPermissionCrud extends AMySqlQueryCrud implements IPermissionCrud
+class MySqlPermissionCrud extends AMySqlPermissionCrud implements IPermissionCrud
 {
-    /**
-     * @var string Name of the table this class performs CRUD operations on.
-     */
-    const PERMISSIONS_TABLE_NAME = 'permissions';
-
-    /**
-     * @var IActionCrud Injected instance of an IActionCrud implementation.
-     */
-    private $actionCrud;
-
-    /**
-     * MySqlPermissionCrud constructor. Injects the MySqlQuery instance used for CRUD operations on
-     * permission data. Injects the IActionCrud implementation instance used for CRUD operations on
-     * action data.
-     * @param MySqlQuery $mySqlQuery
-     * @param IActionCrud $actionCrud
-     */
-    public function __construct(MySqlQuery $mySqlQuery, IActionCrud $actionCrud)
-    {
-        parent::__construct($mySqlQuery, self::PERMISSIONS_TABLE_NAME);
-        $this->actionCrud = $actionCrud;
-    }
-
-    /**
-     * Creates the permissions table.
-     * @return bool True if permissions table was created, false otherwise.
-     */
-    protected function generateTable(): bool
-    {
-        if ($this->MySqlQuery->executeQuery('CREATE TABLE ' . $this->tableName . ' (
-            tableId INT NOT NULL AUTO_INCREMENT PRIMARY KEY UNIQUE,
-            permissionName VARCHAR(242) NOT NULL UNIQUE,
-            permissionActions TEXT NOT NULL,
-            IPermissionType VARCHAR(242) NOT NULL
-        );') === false) {
-            error_log('Permission Crud Error: Failed to create ' . $this->tableName . ' table');
-        }
-        return $this->tableExists($this->tableName);
-    }
 
     /**
      * Create a new permission.
@@ -129,8 +86,12 @@ class MySqlPermissionCrud extends AMySqlQueryCrud implements IPermissionCrud
      */
     public function update(string $permissionName, IPermission $newPermission): bool
     {
-        if ($this->delete($permissionName) === true) {
-            if ($this->create($newPermission) === true) {
+        if ($this->permissionExists($permissionName) === true) {
+            $this->modType = self::MOD_TYPE_UPDATE;
+            $this->originalPermission = $this->read($permissionName);
+            $this->modifiedPermission = $newPermission;
+            if ($this->delete($permissionName) === true && $this->create($newPermission) === true) {
+                $this->notify();
                 return true;
             }
         }
@@ -144,70 +105,19 @@ class MySqlPermissionCrud extends AMySqlQueryCrud implements IPermissionCrud
      */
     public function delete(string $permissionName): bool
     {
-        $this->MySqlQuery->executeQuery('DELETE FROM permissions WHERE permissionName=? LIMIT 1', [$permissionName]);
-        return $this->permissionExists($permissionName) === false;
-    }
-
-    /**
-     * Determine whether or not a specified permission exists.
-     * @param string $permissionName The name of the permission to check for.
-     * @return bool True if permission exists, false otherwise.
-     */
-    private function permissionExists(string $permissionName): bool
-    {
-        $permissionData = $this->MySqlQuery->executeQuery('SELECT * FROM ' . $this->tableName . ' WHERE permissionName=?', [$permissionName])->fetchAll();
-        if (empty($permissionData) === true) {
-            return false;
+        if ($this->permissionExists($permissionName)) {
+            if ($this->modType !== self::MOD_TYPE_UPDATE) {
+                $this->modType = self::MOD_TYPE_DELETE;
+                $this->originalPermission = $this->read($permissionName);
+                $this->modifiedPermission = $this->originalPermission;
+            }
+            $this->MySqlQuery->executeQuery('DELETE FROM permissions WHERE permissionName=? LIMIT 1', [$permissionName]);
+            if ($this->permissionExists($permissionName) === false) {
+                $this->notify();
+                return true;
+            }
         }
-        return true;
-    }
-
-    /**
-     * Get the fully qualified namespaced classname of the specified permission.
-     * @param string $permissionName The name of the permission.
-     * @return string The fully qualified namespaced classname of the specified permission.
-     */
-    private function getClassName(string $permissionName): string
-    {
-        return $this->MySqlQuery->executeQuery('SELECT IPermissionType FROM permissions WHERE permissionName=? LIMIT 1', [$permissionName])->fetchAll(\PDO::FETCH_ASSOC)[0]['IPermissionType'];
-    }
-
-    /**
-     * Creates an array of IAction implementation names based on the IAction implementations
-     * assigned to the IPermission implementation instance, and encodes the generated array as json.
-     * For Example:
-     * {
-     *     [
-     *         'ActionName1',
-     *         'ActionName2',
-     *         'ActionName3',
-     *     ]
-     * }
-     * @param IPermission $Permission The IPermission implementation instance whose Permissions are to be packed.
-     * @return string The packed permission names, i.e., the json string representing the data.
-     */
-    final private function packActions(IPermission $Permission): string
-    {
-        $permissions = array();
-        foreach ($Permission->getActions() as $action) {
-            array_push($permissions, $action->getActionName());
-        }
-        return json_encode($permissions);
-    }
-
-    /**
-     * Unpack the Permission's permissions.
-     * @param string $packedPermissions The packed permissions, i.e., the json string representing the data.
-     * @return array An array of the IPermission implementation instances assigned to the Permission.
-     */
-    final private function unpackActions(string $packedPermissions): array
-    {
-        $unpackedData = json_decode($packedPermissions);
-        $permissions = [];
-        foreach ($unpackedData as $actionName) {
-            array_push($permissions, $this->actionCrud->read($actionName));
-        }
-        return $permissions;
+        return false;
     }
 
 }
